+++ xpumd/receiver/intelxpu/sysman/internal/metadata/generated_metrics_test.go
@@ -67,6 +67,7 @@ func TestMetricsBuilder(t *testing.T) {
 			aggMap["hw.frequency.throttle_status"] = mb.metricHwFrequencyThrottleStatus.config.AggregationStrategy
 			aggMap["hw.gpu.bandwidth.limit"] = mb.metricHwGpuBandwidthLimit.config.AggregationStrategy
 			aggMap["hw.gpu.bandwidth.utilization"] = mb.metricHwGpuBandwidthUtilization.config.AggregationStrategy
+			aggMap["hw.gpu.ecc.state"] = mb.metricHwGpuEccState.config.AggregationStrategy
 			aggMap["hw.gpu.info"] = mb.metricHwGpuInfo.config.AggregationStrategy
 			aggMap["hw.gpu.io"] = mb.metricHwGpuIo.config.AggregationStrategy
 			aggMap["hw.gpu.io.rate"] = mb.metricHwGpuIoRate.config.AggregationStrategy
@@ -144,6 +145,12 @@ func TestMetricsBuilder(t *testing.T) {
 			mb.RecordHwGpuBandwidthUtilizationDataPoint(ts, 1, "hw.id-val", "hw.name-val", "pci.bdf-val")
 			if tt.name == "reaggregate_set" {
 				mb.RecordHwGpuBandwidthUtilizationDataPoint(ts, 3, "hw.id-val-2", "hw.name-val-2", "pci.bdf-val-2")
+			}
+			defaultMetricsCount++
+			allMetricsCount++
+			mb.RecordHwGpuEccStateDataPoint(ts, 1, "hw.id-val", "hw.name-val", "pci.bdf-val", "com.intel.subdevice_id-val", "hw.state-val")
+			if tt.name == "reaggregate_set" {
+				mb.RecordHwGpuEccStateDataPoint(ts, 3, "hw.id-val-2", "hw.name-val-2", "pci.bdf-val-2", "com.intel.subdevice_id-val-2", "hw.state-val-2")
 			}
 			defaultMetricsCount++
 			allMetricsCount++
@@ -254,6 +261,7 @@ func TestMetricsBuilder(t *testing.T) {
 				assert.Empty(t, mb.metricHwFrequencyThrottleStatus.aggDataPoints)
 				assert.Empty(t, mb.metricHwGpuBandwidthLimit.aggDataPoints)
 				assert.Empty(t, mb.metricHwGpuBandwidthUtilization.aggDataPoints)
+				assert.Empty(t, mb.metricHwGpuEccState.aggDataPoints)
 				assert.Empty(t, mb.metricHwGpuInfo.aggDataPoints)
 				assert.Empty(t, mb.metricHwGpuIo.aggDataPoints)
 				assert.Empty(t, mb.metricHwGpuIoRate.aggDataPoints)
@@ -861,6 +869,70 @@ func TestMetricsBuilder(t *testing.T) {
 						_, ok = dp.Attributes().Get("hw.name")
 						assert.False(t, ok)
 						_, ok = dp.Attributes().Get("pci.bdf")
+						assert.False(t, ok)
+					}
+				case "hw.gpu.ecc.state":
+					if tt.name != "reaggregate_set" {
+						assert.False(t, validatedMetrics["hw.gpu.ecc.state"], "Found a duplicate in the metrics slice: hw.gpu.ecc.state")
+						validatedMetrics["hw.gpu.ecc.state"] = true
+						assert.Equal(t, pmetric.MetricTypeSum, mi.Type())
+						assert.Equal(t, 1, mi.Sum().DataPoints().Len())
+						assert.Equal(t, "Known device memory ECC states. Reported only if the ECC state is configurable, or if it has changed.", mi.Description())
+						assert.Equal(t, "1", mi.Unit())
+						assert.False(t, mi.Sum().IsMonotonic())
+						assert.Equal(t, pmetric.AggregationTemporalityCumulative, mi.Sum().AggregationTemporality())
+						dp := mi.Sum().DataPoints().At(0)
+						assert.Equal(t, start, dp.StartTimestamp())
+						assert.Equal(t, ts, dp.Timestamp())
+						assert.Equal(t, pmetric.NumberDataPointValueTypeInt, dp.ValueType())
+						assert.Equal(t, int64(1), dp.IntValue())
+						hwIDAttrVal, ok := dp.Attributes().Get("hw.id")
+						assert.True(t, ok)
+						assert.Equal(t, "hw.id-val", hwIDAttrVal.Str())
+						hwNameAttrVal, ok := dp.Attributes().Get("hw.name")
+						assert.True(t, ok)
+						assert.Equal(t, "hw.name-val", hwNameAttrVal.Str())
+						pciBdfAttrVal, ok := dp.Attributes().Get("pci.bdf")
+						assert.True(t, ok)
+						assert.Equal(t, "pci.bdf-val", pciBdfAttrVal.Str())
+						comIntelSubdeviceIDAttrVal, ok := dp.Attributes().Get("com.intel.subdevice_id")
+						assert.True(t, ok)
+						assert.Equal(t, "com.intel.subdevice_id-val", comIntelSubdeviceIDAttrVal.Str())
+						hwStateAttrVal, ok := dp.Attributes().Get("hw.state")
+						assert.True(t, ok)
+						assert.Equal(t, "hw.state-val", hwStateAttrVal.Str())
+					} else {
+						assert.False(t, validatedMetrics["hw.gpu.ecc.state"], "Found a duplicate in the metrics slice: hw.gpu.ecc.state")
+						validatedMetrics["hw.gpu.ecc.state"] = true
+						assert.Equal(t, pmetric.MetricTypeSum, mi.Type())
+						assert.Equal(t, 1, mi.Sum().DataPoints().Len())
+						assert.Equal(t, "Known device memory ECC states. Reported only if the ECC state is configurable, or if it has changed.", mi.Description())
+						assert.Equal(t, "1", mi.Unit())
+						assert.False(t, mi.Sum().IsMonotonic())
+						assert.Equal(t, pmetric.AggregationTemporalityCumulative, mi.Sum().AggregationTemporality())
+						dp := mi.Sum().DataPoints().At(0)
+						assert.Equal(t, start, dp.StartTimestamp())
+						assert.Equal(t, ts, dp.Timestamp())
+						assert.Equal(t, pmetric.NumberDataPointValueTypeInt, dp.ValueType())
+						switch aggMap["hw.gpu.ecc.state"] {
+						case "sum":
+							assert.Equal(t, int64(4), dp.IntValue())
+						case "avg":
+							assert.Equal(t, int64(2), dp.IntValue())
+						case "min":
+							assert.Equal(t, int64(1), dp.IntValue())
+						case "max":
+							assert.Equal(t, int64(3), dp.IntValue())
+						}
+						_, ok := dp.Attributes().Get("hw.id")
+						assert.False(t, ok)
+						_, ok = dp.Attributes().Get("hw.name")
+						assert.False(t, ok)
+						_, ok = dp.Attributes().Get("pci.bdf")
+						assert.False(t, ok)
+						_, ok = dp.Attributes().Get("com.intel.subdevice_id")
+						assert.False(t, ok)
+						_, ok = dp.Attributes().Get("hw.state")
 						assert.False(t, ok)
 					}
 				case "hw.gpu.info":